python/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grant-os"
version = "0.1.0"
description = "Grant intelligence: Firecrawl scraping, LlamaParse ingestion, compliance matrix, Claude drafting, ReportLab PDFs, LangGraph orchestration."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.49.0",
  "firecrawl-py>=4.22.0",
  "httpx>=0.28.0",
  "langchain-anthropic>=0.3.0",
  "langchain-core>=0.3.0",
  "langgraph>=0.2.0",
  "openai>=1.0.0",
  "pydantic>=2.0.0",
  "pypdf>=5.0.0",
  "python-dotenv>=1.0.0",
  "reportlab>=4.0.0",
]

[project.optional-dependencies]
llama = ["llama-parse>=0.6.0"]
dev = ["ruff>=0.8.0", "pytest>=8.0.0"]

[project.scripts]
grant-scrape = "grant_os.cli:main_scrape"
grant-llamaparse = "grant_os.cli:main_llamaparse"
grant-compliance = "grant_os.cli:main_compliance"
grant-draft = "grant_os.cli:main_draft"
grant-remediate = "grant_os.cli:main_remediate"
grant-pdf = "grant_os.cli:main_pdf"
grant-graph = "grant_os.cli:main_langgraph"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
